requirement.txt
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.4.0
requests==2.28.2
soupsieve==2.3.2.post1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.12.1
